web-animations/animation-player/animation-player-start-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player startTime attribute test</title>
<meta name="assert" content="Returns the start time of this player. Setting this attribute updates the player start time using the procedure defined in section 3.5.2.6 Updating the player start time.">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-startTime">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
function newAnimationPlayer(animationTarget) {
    return document.timeline.play(new Animation(animationTarget, [], 1000));
}

var EPSILON = 10;

test(function() {
    var player = newAnimationPlayer(createDiv(this));

    assert_not_equals(player.startTime, null,
        'The value of AnimationPlayer.startTime attribute should not be of type null');
}, 'The value of AnimationPlayer.startTime attribute is not null');


test(function() {
    var player = newAnimationPlayer(createDiv(this));
    assert_approx_equals(player.startTime, player.timeline.currentTime, EPSILON,
        'Default value of AnimationPlayer.startTime attribute should be the current time of ' +
        'the player timeline');
}, 'Default value of AnimationPlayer.startTime attribute is the current time of ' +
    'the player timeline');


test(function() {
    var player = newAnimationPlayer(createDiv(this));

    var newStartTime = player.startTime + 100;
    player.startTime = newStartTime;

    assert_equals(player.startTime, newStartTime,
        'The value of AnimationPlayer.startTime attribute should be set to the new value');
	// TODO Is there a way to test the following "2. Set the hold time to null."?
}, 'Test setting value of AnimationPlayer.startTime attribute. Player is not paused');


test(function() {
    var player = newAnimationPlayer(createDiv(this));

    var newStartTime = player.startTime + 100;
    player.pause();
    player.startTime = newStartTime;

    var expectedTime = player.timeline.currentTime - newStartTime * player.playbackRate;
    assert_equals(player.startTime, expectedTime,
        'The value of AnimationPlayer.startTime attribute should be set to ' +
        'hod time value');
}, 'Test setting value of AnimationPlayer.startTime attribute. Player is paused');


async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.startTime += 50;

    setTimeout(function() {
        t.step(function () {
            assert_approx_equals(player.currentTime, 50, EPSILON, 'The value of ' +
                'AnimationPlayer.currentTime should correspond startTime');
            t.done();
        });
    }, 100);
}, 'Test setting AnimationPlayer start time delays beginning of play');


async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.startTime -= 100;

    setTimeout(function() {
        t.step(function () {
            assert_approx_equals(player.currentTime, 150, EPSILON, 'The value of ' +
                'AnimationPlayer.currentTime should correspond startTime');
            t.done();
        });
    }, 50);
}, 'Test setting AnimationPlayer start time shifts beginning of play (start time ' +
    'is less than current time)');


async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.pause();
    player.startTime += 50;

    setTimeout(function() {
        t.step(function () {
            assert_equals(player.currentTime, 0, 'The value of ' +
                'AnimationPlayer.currentTime should correspond startTime');
            t.done();
        });
    }, 60);
}, 'Test setting AnimationPlayer start time delays beginning of play (player is paused)');


async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.pause();
    player.startTime -= 50;

    setTimeout(function() {
        t.step(function () {
            assert_equals(player.currentTime, 0, 'The value of ' +
                'AnimationPlayer.currentTime should correspond startTime');
            t.done();
        });
    }, 60);
}, 'Test setting AnimationPlayer start time shifts beginning of play (player is paused, ' +
    'start tima less than current time)');
</script>
</body>
